Phase 2: Modern project restructuring
- Create modern Go project structure: cmd/, internal/, pkg/
- Move main entry point to cmd/eseries_exporter/main.go
- Reorganize collectors into internal/collectors/
- Reorganize config into internal/config/
- Update all imports to use new module paths
- Move test file to correct location (cmd/eseries_exporter/)
- Fix test data file paths
- Ensure successful compilation
- Apply modern Go project layout conventions
