+++ internal/collectors/controller-statistics_test.go
@@ -32,38 +32,38 @@ import (
 
 	"github.com/go-kit/log"
 	"github.com/prometheus/client_golang/prometheus/testutil"
-	"github.com/sckyzo/eseries_exporter/config"
+	"github.com/sckyzo/eseries_exporter/internal/config"
 )
 
-func TestDriveStatisticsCollector(t *testing.T) {
-	analyzedDriveData, err := os.ReadFile("testdata/analysed-drive-statistics.json")
+func TestControllerStatisticsCollector(t *testing.T) {
+	analyzedControllerData, err := os.ReadFile("testdata/analysed-controller-statistics.json")
 	if err != nil {
 		t.Fatalf("Error loading fixture data: %s", err.Error())
 	}
-	driveData, err := os.ReadFile("testdata/drive-statistics.json")
+	controllerData, err := os.ReadFile("testdata/controller-statistics.json")
 	if err != nil {
 		t.Fatalf("Error loading fixture data: %s", err.Error())
 	}
-	inventoryData, err := os.ReadFile("testdata/drives.json")
+	inventoryData, err := os.ReadFile("testdata/controllers.json")
 	if err != nil {
 		t.Fatalf("Error loading fixture data: %s", err.Error())
 	}
 	expected := `
-	# HELP eseries_drive_average_read_op_size_bytes Drive statistic averageReadOpSize
-	# TYPE eseries_drive_average_read_op_size_bytes gauge
-	eseries_drive_average_read_op_size_bytes{slot="58",tray="0"} 39620.99569760295
-	eseries_drive_average_read_op_size_bytes{slot="53",tray="0"} 21312.646464646463
+	# HELP eseries_controller_average_read_op_size_bytes Controller statistic averageReadOpSize
+	# TYPE eseries_controller_average_read_op_size_bytes gauge
+	eseries_controller_average_read_op_size_bytes{controller="070000000000000000000001",controller_label="A"} 39687.27392305163
+	eseries_controller_average_read_op_size_bytes{controller="070000000000000000000002",controller_label="B"} 73664.54585344449
 	# HELP eseries_exporter_collect_error Indicates if error has occurred during collection
 	# TYPE eseries_exporter_collect_error gauge
-	eseries_exporter_collect_error{collector="drive-statistics"} 0
+	eseries_exporter_collect_error{collector="controller-statistics"} 0
 	`
 	server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
 		if strings.HasSuffix(req.URL.Path, "hardware-inventory") {
 			_, _ = rw.Write(inventoryData)
-		} else if strings.HasSuffix(req.URL.Path, "analysed-drive-statistics") {
-			_, _ = rw.Write(analyzedDriveData)
+		} else if strings.HasSuffix(req.URL.Path, "analysed-controller-statistics") {
+			_, _ = rw.Write(analyzedControllerData)
 		} else {
-			_, _ = rw.Write(driveData)
+			_, _ = rw.Write(controllerData)
 		}
 	}))
 	defer server.Close()
@@ -77,33 +77,25 @@ func TestDriveStatisticsCollector(t *testing.T) {
 	}
 	w := log.NewSyncWriter(os.Stderr)
 	logger := log.NewLogfmtLogger(w)
-	collector := NewDriveStatisticsExporter(target, logger)
+	collector := NewControllerStatisticsExporter(target, logger)
 	gatherers := setupGatherer(collector)
 	if val, err := testutil.GatherAndCount(gatherers); err != nil {
 		t.Errorf("Unexpected error: %v", err)
-	} else if val != 48 {
-		t.Errorf("Unexpected collection count %d, expected 48", val)
+	} else if val != 56 {
+		t.Errorf("Unexpected collection count %d, expected 56", val)
 	}
 	if err := testutil.GatherAndCompare(gatherers, strings.NewReader(expected),
-		/*
-			"eseries_drive_average_read_op_size", "eseries_drive_average_write_op_size",
-			"eseries_drive_combined_iops", "eseries_drive_combined_response_time", "eseries_drive_combined_throughput",
-			"eseries_drive_read_iops", "eseries_drive_read_ops", "eseries_drive_read_physical_iops",
-			"eseries_drive_read_response_time", "eseries_drive_read_throughput",
-			"eseries_drive_write_iops", "eseries_drive_write_ops", "eseries_drive_write_physical_iops",
-			"eseries_drive_write_response_time", "eseries_drive_write_throughput",
-		*/
-		"eseries_drive_average_read_op_size_bytes",
+		"eseries_controller_average_read_op_size_bytes",
 		"eseries_exporter_collect_error"); err != nil {
 		t.Errorf("unexpected collecting result:\n%s", err)
 	}
 }
 
-func TestDriveStatisticsCollectorError(t *testing.T) {
+func TestControllerStatisticsCollectorError(t *testing.T) {
 	expected := `
 	# HELP eseries_exporter_collect_error Indicates if error has occurred during collection
 	# TYPE eseries_exporter_collect_error gauge
-	eseries_exporter_collect_error{collector="drive-statistics"} 1
+	eseries_exporter_collect_error{collector="controller-statistics"} 1
 	`
 	server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
 		http.Error(rw, "error", http.StatusNotFound)
@@ -119,7 +111,7 @@ func TestDriveStatisticsCollectorError(t *testing.T) {
 	}
 	w := log.NewSyncWriter(os.Stderr)
 	logger := log.NewLogfmtLogger(w)
-	collector := NewDriveStatisticsExporter(target, logger)
+	collector := NewControllerStatisticsExporter(target, logger)
 	gatherers := setupGatherer(collector)
 	if val, err := testutil.GatherAndCount(gatherers); err != nil {
 		t.Errorf("Unexpected error: %v", err)
@@ -127,7 +119,7 @@ func TestDriveStatisticsCollectorError(t *testing.T) {
 		t.Errorf("Unexpected collection count %d, expected 2", val)
 	}
 	if err := testutil.GatherAndCompare(gatherers, strings.NewReader(expected),
-		"eseries_drive_average_read_op_size_bytes", "eseries_exporter_collect_error"); err != nil {
+		"eseries_controller_average_read_op_size_bytes", "eseries_exporter_collect_error"); err != nil {
 		t.Errorf("unexpected collecting result:\n%s", err)
 	}
 }
