+++ internal/collectors/drive-statistics_test.go
@@ -32,24 +32,39 @@ import (
 
 	"github.com/go-kit/log"
 	"github.com/prometheus/client_golang/prometheus/testutil"
-	"github.com/sckyzo/eseries_exporter/config"
+	"github.com/sckyzo/eseries_exporter/internal/config"
 )
 
-func TestSystemStatisticsCollector(t *testing.T) {
-	fixtureData, err := os.ReadFile("testdata/system-statistics.json")
+func TestDriveStatisticsCollector(t *testing.T) {
+	analyzedDriveData, err := os.ReadFile("testdata/analysed-drive-statistics.json")
+	if err != nil {
+		t.Fatalf("Error loading fixture data: %s", err.Error())
+	}
+	driveData, err := os.ReadFile("testdata/drive-statistics.json")
+	if err != nil {
+		t.Fatalf("Error loading fixture data: %s", err.Error())
+	}
+	inventoryData, err := os.ReadFile("testdata/drives.json")
 	if err != nil {
 		t.Fatalf("Error loading fixture data: %s", err.Error())
 	}
 	expected := `
+	# HELP eseries_drive_average_read_op_size_bytes Drive statistic averageReadOpSize
+	# TYPE eseries_drive_average_read_op_size_bytes gauge
+	eseries_drive_average_read_op_size_bytes{slot="58",tray="0"} 39620.99569760295
+	eseries_drive_average_read_op_size_bytes{slot="53",tray="0"} 21312.646464646463
 	# HELP eseries_exporter_collect_error Indicates if error has occurred during collection
 	# TYPE eseries_exporter_collect_error gauge
-	eseries_exporter_collect_error{collector="system-statistics"} 0
-	# HELP eseries_system_average_read_op_size_bytes System statistic averageReadOpSize
-	# TYPE eseries_system_average_read_op_size_bytes gauge
-	eseries_system_average_read_op_size_bytes 17357.11013434037
+	eseries_exporter_collect_error{collector="drive-statistics"} 0
 	`
 	server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
-		_, _ = rw.Write(fixtureData)
+		if strings.HasSuffix(req.URL.Path, "hardware-inventory") {
+			_, _ = rw.Write(inventoryData)
+		} else if strings.HasSuffix(req.URL.Path, "analysed-drive-statistics") {
+			_, _ = rw.Write(analyzedDriveData)
+		} else {
+			_, _ = rw.Write(driveData)
+		}
 	}))
 	defer server.Close()
 	baseURL, _ := url.Parse(server.URL)
@@ -62,33 +77,33 @@ func TestSystemStatisticsCollector(t *testing.T) {
 	}
 	w := log.NewSyncWriter(os.Stderr)
 	logger := log.NewLogfmtLogger(w)
-	collector := NewSystemStatisticsExporter(target, logger)
+	collector := NewDriveStatisticsExporter(target, logger)
 	gatherers := setupGatherer(collector)
 	if val, err := testutil.GatherAndCount(gatherers); err != nil {
 		t.Errorf("Unexpected error: %v", err)
-	} else if val != 14 {
-		t.Errorf("Unexpected collection count %d, expected 14", val)
+	} else if val != 48 {
+		t.Errorf("Unexpected collection count %d, expected 48", val)
 	}
 	if err := testutil.GatherAndCompare(gatherers, strings.NewReader(expected),
 		/*
-			"eseries_system_average_read_op_size", "eseries_system_average_write_op_size",
-			"eseries_system_combined_iops", "eseries_system_combined_response_time", "eseries_system_combined_throughput",
-			"eseries_system_read_iops", "eseries_system_read_ops", "eseries_system_read_physical_iops",
-			"eseries_system_read_response_time", "eseries_system_read_throughput",
-			"eseries_system_write_iops", "eseries_system_write_ops", "eseries_system_write_physical_iops",
-			"eseries_system_write_response_time", "eseries_system_write_throughput",
+			"eseries_drive_average_read_op_size", "eseries_drive_average_write_op_size",
+			"eseries_drive_combined_iops", "eseries_drive_combined_response_time", "eseries_drive_combined_throughput",
+			"eseries_drive_read_iops", "eseries_drive_read_ops", "eseries_drive_read_physical_iops",
+			"eseries_drive_read_response_time", "eseries_drive_read_throughput",
+			"eseries_drive_write_iops", "eseries_drive_write_ops", "eseries_drive_write_physical_iops",
+			"eseries_drive_write_response_time", "eseries_drive_write_throughput",
 		*/
-		"eseries_system_average_read_op_size_bytes",
+		"eseries_drive_average_read_op_size_bytes",
 		"eseries_exporter_collect_error"); err != nil {
 		t.Errorf("unexpected collecting result:\n%s", err)
 	}
 }
 
-func TestSystemStatisticsCollectorError(t *testing.T) {
+func TestDriveStatisticsCollectorError(t *testing.T) {
 	expected := `
 	# HELP eseries_exporter_collect_error Indicates if error has occurred during collection
 	# TYPE eseries_exporter_collect_error gauge
-	eseries_exporter_collect_error{collector="system-statistics"} 1
+	eseries_exporter_collect_error{collector="drive-statistics"} 1
 	`
 	server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
 		http.Error(rw, "error", http.StatusNotFound)
@@ -104,7 +119,7 @@ func TestSystemStatisticsCollectorError(t *testing.T) {
 	}
 	w := log.NewSyncWriter(os.Stderr)
 	logger := log.NewLogfmtLogger(w)
-	collector := NewSystemStatisticsExporter(target, logger)
+	collector := NewDriveStatisticsExporter(target, logger)
 	gatherers := setupGatherer(collector)
 	if val, err := testutil.GatherAndCount(gatherers); err != nil {
 		t.Errorf("Unexpected error: %v", err)
@@ -112,7 +127,7 @@ func TestSystemStatisticsCollectorError(t *testing.T) {
 		t.Errorf("Unexpected collection count %d, expected 2", val)
 	}
 	if err := testutil.GatherAndCompare(gatherers, strings.NewReader(expected),
-		"eseries_system_average_read_op_size_bytes", "eseries_exporter_collect_error"); err != nil {
+		"eseries_drive_average_read_op_size_bytes", "eseries_exporter_collect_error"); err != nil {
 		t.Errorf("unexpected collecting result:\n%s", err)
 	}
 }
