+++ internal/collectors/testdata/analysed-controller-statistics.json
