+++ internal/collectors/testdata/controller-statistics.json
