+++ internal/collectors/testdata/drive-statistics.json
