+++ internal/collectors/testdata/drives-duplicate.json
