+++ internal/collectors/testdata/hardware-inventory.json
