+++ internal/collectors/testdata/rootCA.key
