+++ internal/collectors/testdata/storage-systems.json
