+++ internal/collectors/testdata/system-statistics.json
