+++ internal/config/testdata/eseries_exporter.yaml
