+++ internal/config/testdata/missing-proxy-url.yaml
